+++ src/main/java/com/elmakers/mine/bukkit/plugins/magic/CleanupBlocksTask.java
@@ -5,14 +5,16 @@ import com.elmakers.mine.bukkit.dao.BlockList;
 public class CleanupBlocksTask implements Runnable
 {
 	protected BlockList undoBlocks;
+	protected Spells spells;
 
-	public CleanupBlocksTask(BlockList cleanup)
+	public CleanupBlocksTask(Spells spells, BlockList cleanup)
 	{
 		this.undoBlocks = cleanup;
+		this.spells = spells;
 	}
 
 	public void run()
 	{
-		this.undoBlocks.undo();
+		this.undoBlocks.undo(spells);
 	}
 }